Module kaleidoscope_lexer::token

Types which represent a token in a Kaleidoscope program. These structures describe the various aspects a token has, including the kind of token it is and where it is found.

Modules

bracket

A bunch of types representing a bracket such as ( or ]. Objects of these types can also check to see if they match or cancel out each other, which can make sure that the brackets in a programme are balanced out properly.

brackets

All the possible bracket types that can be represented by the Bracket struct.

fileindex

A representation of the location of a character in a file. This can aid in identifying the source of a syntax error or which part of the code caused or encountered an error at compile time or runtime.

keyword

A enum representing a keyword. A key reason why I chose to define a special enum because manually checking if the span in a token is what I was looking for matches a keyword.

operator

A token representing an operator.

token

A token in a file or stream. This is the most basic unit in a file above a character and includes integers, floats, names and operators.

tokenkind

The type of the token. Such a classification allows the parser to make sure that the input given is correct.

Structs

Bracket

A struct representing a bracket. This bracket can be grouped by its kind and side (see BracketKind and BracketSide respectively).

FileIndex

Represents the location of a character (i.e. displayed glyphs, diacritics are counted as separate symbols) in a file or stream.

Token

A token in a Kaleidoscope file.

Enums

BracketKind

The type of bracket Bracket represents. This merely classifies a bracket by its shape (i.e. whether it’s round, square or curly, etc.) and not its orientation.

BracketSide

The orientation or side of a bracket. This can either be left or right. Brackets which are classified as “Left” include “(”, “[” and “{”, etc, with their corresponding counterparts (“)”, “]”, “}” respectively) are counted as “Right”-side brackets.

Keyword

An enumerator of possible keywords that can be encountered in Kaleidoscope.

Operator

A token representing an operator.

TokenKind

The type of token a token is.

Constants

LEFT_ANGLED_BRACKET

<

LEFT_CURLY_BRACKET

{

LEFT_ROUND_BRACKET

(

LEFT_SQUARE_BRACKET

[

RIGHT_ANGLED_BRACKET

>

RIGHT_CURLY_BRACKET

}

RIGHT_ROUND_BRACKET

)

RIGHT_SQUARE_BRACKET

]

UNKNOWN_BRACKET

???