Module kaleidoscope_lexer::token::bracket

A bunch of types representing a bracket such as ( or ]. Objects of these types can also check to see if they match or cancel out each other, which can make sure that the brackets in a programme are balanced out properly.

Modules

brackets
All the possible bracket types that can be represented by the Bracket struct.

Structs

Bracket
A struct representing a bracket. This bracket can be grouped by its kind and side (see BracketKind and BracketSide respectively).

Enums

BracketKind
The type of bracket Bracket represents. This merely classifies a bracket by its shape (i.e. whether it’s round, square or curly, etc.) and not its orientation.

BracketSide
The orientation or side of a bracket. This can either be left or right. Brackets which are classified as “Left” include “(”, “[” and “{”, etc, with their corresponding counterparts (“)”, “]”, “}” respectively) are counted as “Right”-side brackets.