Module kaleidoscope_lexer::utils

Utility functions for recognising tokens and checking whether they have certain “characteristics” that makes them suitable to be part of a token of a certain type.

For example, an identifier must start with an ASCII letter (both miniscule and majuscule) or an underscore _. Any following characters must be an ASCII letter, underscore or numeric digit (0 to 9). If you want to ensure that a given string can be a valid identifier, you can create a function that checks the correctness of the identifier and return the index of the first character that causes the string to fail the test. An example implementation is shown below.

use kaleidoscope_lexer::utils;

fn validate_identifier(string: &str) -> (bool, Option<usize>) {
    if string.len() == 0 {
        (false, None)
    } else {
        let units: Vec<char> = string.chars().collect();
        if !utils::is_identifier_start(units[0]) {
            return (false, Some(0));
        }
        for (index, unit) in units[1..].iter().enumerate() {
            if !utils::is_identifier(*unit) {
                return (false, Some(index + 1));
            }
        }
        (true, None)
    }
}

let (ok, index) = validate_identifier("citizen_839293927392738");
assert!(ok);
assert!(index.is_none());
let (ok, index) = validate_identifier("7huhfe");
assert!(!ok);
assert!(matches!(index, Some(0)));
let (ok, index) = validate_identifier("");
assert!(!ok);
assert!(index.is_none());
let (ok, index) = validate_identifier("hmm???");
assert!(!ok);
assert!(matches!(index, Some(3)));

This is not the only application of the functions defined in this module. You can tests for brackets, operators, numbers and punctuation, which play an important role in the grammar of the language.

Functions

is_alpha
True if unit is an ASCII alphabetical character.

is_alphanum
True if unit is an ASCII alphabetical character or an ASCII decimal digit.

is_bracket
See if unit is a bracket character.

is_comma
See if unit is a comma separator.

is_comment
See if unit is the character that denotes the start of a comment.

is_decimal_digit
True if unit is within ‘0’ and ‘9’.

is_dot
See if unit is a fullstop.

is_eol
Rudimentary check to see if the end of the line has been reached.

is_fullstop
True if unit is ‘.’

is_identifier
True if unit is a character that can potentially be used in an identifier.

is_identifier_start
True if unit is a character that can act as the first character of a name/identifier.

is_opchar
True if unit is a character used in one of the operator symbols.

is_semicolon
See if unit is a semicolon.

is_whitespace
True if unit is a whitespace character, including newlines.